extract live and static message items for separate uses
