+++ src/assets/main.css
@@ -623,6 +623,15 @@ button:has(.channelview-avatar) {
   }
 }
 
+.modal-messages {
+  .event-item {
+    padding: 0.5rem;
+    border-radius: var(--radius);
+    width: 100%;
+    background-color: oklch(from var(--base) l c h / 0.5);
+  }
+}
+
 /* active users */
 
 .activeroll {
